requirements.txt
torch>=2.0
torchvision>=0.15
Pillow>=9.0
torchmetrics[image]>=1.0
datasets>=2.0
tqdm>=4.0
scikit-image>=0.21
